Struct eccodes::codes_handle::KeyedMessage

pub struct KeyedMessage { /* fields omitted */ }

Structure used to access keys inside the GRIB file message. All data (including data values) contained by the file can only be accessed through the message and keys.

The structure implements Clone trait which comes with a memory overhead. You should take care that your system has enough memory before cloning KeyedMessage.

Keys inside the message can be accessed directly with read_key() function or using FallibleIterator. The function find_nearest() allows to get the values of four nearest gridpoints to requested coordinates. FallibleIterator parameters can be set with set_iterator_parameters() to specify the subset of keys to iterate over.

Implementations

impl KeyedMessage

pub fn set_iterator_parameters(
&mut self,
flags: Vec<KeysIteratorFlags>,
namespace: String
)

Function that allows to set the flags and namespace for FallibleIterator. Must be called before calling the iterator. Changing the parameters after first call of next() will have no effect on the iterator.

The flags are set by providing any combination of KeysIteratorFlags inside a vector. Check the documentation for the details of each flag meaning.

Namespace is set simply as string, eg. "ls", "time", "parameter", "geography", "statistics". Invalid namespace will result in empty iterator.

Default parameters are AllKeys flag and "" namespace, which implies iteration over all keys available in the message.

Example

let file_path = Path::new("./data/iceland.grib");
let product_kind = ProductKind::GRIB;

let mut handle = CodesHandle::new_from_file(file_path, product_kind).unwrap();
let mut current_message = handle.next().unwrap().unwrap();

let flags = vec![
    KeysIteratorFlags::AllKeys,
    KeysIteratorFlags::SkipOptional,
    KeysIteratorFlags::SkipReadOnly,
    KeysIteratorFlags::SkipDuplicates,
];

let namespace = "geography".to_owned();

current_message.set_iterator_parameters(flags, namespace);

while let Some(key) = current_message.next().unwrap() {
    println!("{:?}", key);
}

impl KeyedMessage

pub fn read_key(&self, key_name: &str) -> Result<Key, CodesError>

Method to get a Key with provided name from the KeyedMessage.

This function takes a key name and returns the key value as Key if requested key exists. Check the Key documentation for details of possible key types.

Example

let file_path = Path::new("./data/iceland.grib");
let product_kind = ProductKind::GRIB;

let mut handle = CodesHandle::new_from_file(file_path, product_kind).unwrap();
let message = handle.next().unwrap().unwrap();
let message_short_name = message.read_key("shortName").unwrap();

assert_eq!(message_short_name.value, Str("msl".to_string()));

This function will try to retrieve the key of native string type as string even when the nul byte is not positioned at the end of key value.

If retrieving the key value in native type fails this function will try to read the requested key as bytes.

Errors

Returns CodesInternal::CodesNotFound wrapped in CodesError::Internal when a key of given name has not been found in the message.

Returns CodesError::MissingKey when a given key has a missing type.

Returns CodesError::Internal when one of internal ecCodes functions to read the key fails.

Returns CodesError::CstrUTF8 and CodesError::NulChar when the string returned by ecCodes library cannot be parsed as valid UTF8 Rust string.

Panics

Panics when the size of given key is lower than 1. This indicates corrupted data file, bug in the crate or bug in the ecCodes library. If you encounter this panic please check if your file is correct and report it on Github.

impl KeyedMessage

pub fn write_to_file(
&self,
file_path: &Path,
append: bool
) -> Result<(), CodesError>

Function to write given KeyedMessage to a file at provided path. If file does not exists it will be created. If append is set to true file will be opened in append mode and no data will be overwritten (useful when writing mutiple messages to one file).

Example

let in_path = Path::new("./data/iceland-levels.grib");
let out_path  = Path::new("./data/iceland-temperature-levels.grib");

let handle = CodesHandle::new_from_file(in_path, GRIB)?;

let mut t_levels =
    handle.filter(|msg| Ok(msg.read_key("shortName")?.value == Str("t".to_string())));

while let Some(msg) = t_levels.next()? {
    msg.write_to_file(out_path, true)?;
}

Errors

Returns CodesError::FileHandlingInterrupted when the file cannot be opened, created or correctly written.

Returns CodesInternal when internal ecCodes function returns non-zero code.

pub fn write_key(&mut self, key: Key) -> Result<(), CodesError>

Function to set specified Key inside the KeyedMessage. This function automatically matches the KeyType and uses adequate internal ecCodes function to set the key. The message must be mutable to use this function.

User must provide the Key with correct type, otherwise error will occur. Note that not all keys can be set, for example "name" and shortName are read-only. Trying to set such keys will result in error. Some keys can also be set using a non-native type (eg. centre), but read_key() function will only read then in native type.

Refer to ecCodes library documentation for more details.

Example

let file_path = Path::new("./data/iceland.grib");

let mut handle = CodesHandle::new_from_file(file_path, GRIB).unwrap();
let mut current_message = handle.next().unwrap().unwrap();

let new_key = Key {
    name: "centre".to_string(),
    value: KeyType::Str("cnmc".to_string()),
};

current_message.write_key(new_key).unwrap();

Errors

This method will return CodesInternal when internal ecCodes function returns non-zero code.

impl KeyedMessage

pub fn find_nearest(
&mut self,
lat: f64,
lon: f64
) -> Result<[NearestGridpoint; 4], CodesError>

Function to get four NearestGridpoints of a point represented by requested coordinates.

The inputs are latitude and longitude of requested point in respectively degrees north and degreed east.

In the output gridpoints, the value field refers to parameter held by the KeyedMessage for which the function is called in adequate units, coordinates are in degrees north/east, and distance field represents the distance between requested point and output point in kilometers.

Example

let file_path = Path::new("./data/iceland.grib");
let product_kind = ProductKind::GRIB;

let mut handle = CodesHandle::new_from_file(file_path, product_kind).unwrap();
let mut msg = handle.next().unwrap().unwrap();

let out = msg.find_nearest(64.13, -21.89).unwrap();

Errors

This function returns CodesInternal when one of ecCodes function returns the non-zero code.